+++ apps/opentelemetry_api_experimental/lib/open_telemetry/meter.ex
@@ -19,6 +19,11 @@ defmodule OpenTelemetryAPIExperimental.Meter do
   end
 
   defmacro register_callback(instruments, callback, callback_args) do
-    :otel_meter.register_callback(:opentelemetry_experimental.get_meter(__MODULE__), instruments, callback, callback_args)
+    :otel_meter.register_callback(
+      :opentelemetry_experimental.get_meter(__MODULE__),
+      instruments,
+      callback,
+      callback_args
+    )
   end
 end